// src/app/shared/interfaces/auxilium/patient-center/patient-ar-history.interface.ts
export interface PatientArHistory {
    id: number;
    chargeid: number;
    patientid: number;
    invoiceno: string;
    itemid: number;
    patientpayorid: number;
    branchid: number;
    svcdate: Date;
    patalpha: string;
    todate: Date;
    sequence: string;
    hcpcscode: string;
    printstatus: string;
    shipstatus: string;
    trantype: string;
    billtype: string;
    mod1: string;
    mod2: string;
    mod3: string;
    submitted: number;
    allowed: number;
    tranamount: number;
    balance: number;
    createdate: Date;
    printdate: Date;
    origdate: Date;
    quantity: number;
    salesid: number;
    cmnformid: number;
    icn: string;
    billform: string;
    amountpaid: number;
    amountadjusted1: number;
    checkno: string;
    checkdate: Date;
    postdate: Date;
    batchno: string;
    physician: string;
    patientdoctorid1: number;
    seq: string;
    referid: number;
    diagnosis: string;
    patienticdid1: number;
    held: string;
    billcycle: string;
    status: boolean;
    collectionstatus: string;
    statusdate: Date;
    datebilled1: Date;
    datebilled2: Date;
    datebilled3: Date;
    paponumber: string;
    adjustmentcode1: string;
    adjustmentcode2: string;
    amountadjusted2: number;
    adjustmentcode3: string;
    amountadjusted3: number;
    adjustmentcode4: string;
    amountadjusted4: number;
    adjustmentcode5: string;
    amountadjusted5: number;
    adjustmentcode6: string;
    amountadjusted6: number;
    billto: string;
    claimtype: string;
    deletedate: Date;
    deleteuserid: string;
    pwktypecode: string;
    pwktransmissioncode: string;
    pwkattachmentcn: string;
    pwkccn: string;
    lot: boolean;
    upnqualifier: string;
    upn: string;
    level: string;
    patientdoctorid2: number;
    msg1: string;
    msg2: string;
    msg3: string;
    msg4: string;
    msg5: string;
    claimstatus: string;
    paymenttype: string;
    paymenttype2: string;
    priceid: number;
    mod4: string;
    accessionno: string;
    adduserid: string;
    confirmid: string;
    pos: string;
    isincollections: boolean;
    servicedate: Date;
    postingdate: Date;
    epsdt: string;
    ernpostingfile: string;
    itemcode?: string;
    description?: string;
}